server1/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
boto3==1.35.24
celery==5.4.0
redis==5.0.8
python-multipart==0.0.9
numpy==1.26.4
opencv-python==4.10.0.84
mediapipe==0.10.14
torch==2.4.1
torchvision==0.19.1
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
loguru==0.7.2